// src/app/components/types.tsx
export interface PersonData {
    id: string;
    name: string;
    age: number;
    avatar: string;
    pontos: number;
    dataCriacao: string;
    fasesCompletas: {
      compareExpressions: {
        facil: number;
        medio: number;
        dificil: number;
      };
      escrevaExpressao: {
        facil: number;
        medio: number;
        dificil: number;
      };
      selecioneExpressao: {
        facil: number;
        medio: number;
        dificil: number;
      };
      ligueExpressao: {
        facil: number;
        medio: number;
        dificil: number;
      };
      encontreOErro: {
        facil: number;
        medio: number;
        dificil: number;
      };
      jogoDaMemoria: {
        facil: number;
        medio: number;
        dificil: number;
      };
      piscaPica: {
        facil: number;
        medio: number;
        dificil: number;
      };
      sorriaEstaSendoFilmado: {
        facil: number;
        medio: number;
        dificil: number;
      };
      chefinMandou: {
        facil: number;
        medio: number;
        dificil: number;
      };
      imiteAExpressao: {
        facil: number;
        medio: number;
        dificil: number;
      };
     
    };
    tempoJogado: {
      compareExpressions: {
        facil: number;
        medio: number;
        dificil: number;
      };
      escrevaExpressao: {
        facil: number;
        medio: number;
        dificil: number;
      };
      selecioneExpressao: {
        facil: number;
        medio: number;
        dificil: number;
      };
      ligueExpressao: {
        facil: number;
        medio: number;
        dificil: number;
      };
      encontreOErro: {
        facil: number;
        medio: number;
        dificil: number;
      };
      jogoDaMemoria: {
        facil: number;
        medio: number;
        dificil: number;
      };
      piscaPica: {
        facil: number;
        medio: number;
        dificil: number;
      };
      sorriaEstaSendoFilmado: {
        facil: number;
        medio: number;
        dificil: number;
      };
      chefinMandou: {
        facil: number;
        medio: number;
        dificil: number;
      };
      imiteAExpressao: {
        facil: number;
        medio: number;
        dificil: number;
      };
     
    };

  }

export interface Token {
    hash: string
}